cdba: Make exit code reflect exit cause
Make the exit code non-zero when the cdba for some reason failed, to
allow tools calling cdba to differentiate the various scenarios.

There is still some tuning needed here, but for now we exit with error
code 110 if we reached a timeout, 1 if there was an error and 0 if we
either exited by user request or by hitting the power-off token.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>