cdba: Add -C argument for power-cycle on success only
When using cdba to bisect for intermittent issues related to boot
lockups or crashes taking us to download mode it's only useful to
power-cycle on success (as a failure case shows the intermittent
problem).

Add a special version of the existing power cycle code, that stops the
power cycling loop if a timeout criteria is hit, cutting down on
unnecessary reboots of a broken kernel.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>