cdba: Don't default to fastboot endlessly
The current scheme of endlessly rebooting means that we spend -t amount
of time to boot the device over and over if the device crashes early.

Add a new flag -R which does this, but change the default so that if we
see the device showing up again in fastboot we consider it rebooted and
quits.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>