cdba: Revamp timeout handling to support both -t and -T
Revamp the implementation of timeouts, to allow for both a total timeout
(-t) and an inactivity timeout -T at the same time. This is useful for
CI-loops where a board ending up in download mode should be
released quickly (due to inactivity), but a full test can be rather
extensive.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>