cdba.c: Move common code into cdba-client.c
I aim to implement two kinds of cdba clients one to cover developer
workflow interactively (select board, get console, power on, flash/boot and test)
and another to be enable  LAVA automation, cdba-client.c will be the
module that implements the common support for both of it.

A common set of functions are needed to implement by both cdba.c and
cdba-standalone.c,

  void cdba_usage(const char *__progname);
  int cdba_handle_opt(int argc, const char *argv);
  void cdba_handle_verb(int argc, char **argv, int optind, const char *board);
  int cdba_handle_message(struct msg *msg);
  void cdba_end(void);

And a set of helpers exposed by cdba_client can be used, that has common
functionality between this two implementations,

  void cdba_client_usage(void);
  void cdba_client_request_select_board(const char *board);
  void cdba_client_request_power_on(void);
  void cdba_client_request_power_off(void);
  int cdba_client_reached_timeout(void);
  int cdba_client_main(int argc, char **argv, const char *opts);

I plan to implement cdba-standalone.c that extends cdba-client.c
providing support for reserve a board and do a set of actions
asyncronusly when LAVA tries to execute certain job in the board.

Signed-off-by: Aníbal Limón <anibal.limon@linaro.org>