cdba: Add support of standalone client and server
The cdba-standalone is a client that supports to get console and power
on/off the device, the cdba-standalone-server makes usage of a
communication channel to connect to already running cdba-server and
request power on/off.

The power on/off actions finalized succesfully but shown,

...
cdba-standalone: received -1 on stdout: Broken pipe
...

Due to the initial design because always has a console attached
via ssh not useful in power on/off. To avoid this error will need to
write a new cdba-server to cover this workflow better.

Signed-off-by: Aníbal Limón <anibal.limon@linaro.org>