cdba-server: add CDBA-server banner to be output on startup
Granted all possible issues with SSH, especially in the remote and/or
proxied labs, and possible issues with the board being present but
failing to power up, it is useful to have a banner message once CDBA was
started. This gives user a clear picture that the connection is fine and
the server is trying to power up the board.

Signed-off-by: Dmitry Baryshkov <dmitry.baryshkov@linaro.org>