support custom reset sequences
Proper power sequencing with VBATT and VBUS supplies that can be
switched on and off can be tricky, especially for production devices
that use a battery rather than a 12v supply.

Add the ability to define a custom button sequence to reset a device,
this is necessary rather than hard-coding because the exact behaviour
can vary wildly on different devices.

Document this in the README and provide an example.

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>