cdba-server: provide fallbacks for the username value
We use the CDBA_USER value as a way to pass info from cdba-shell to
cdba-server. However if it is not provide (e.g. because the cdba-server
is started directly by shell) add a fallback to the system's USER
variable. And if that's also not set, provide a safe default of
'nobody'.

Signed-off-by: Dmitry Baryshkov <dmitry.baryshkov@linaro.org>